Make clock_t unsigned int everywhere. Ok: matt@, mrg@

+++ sys/arch/m68k/include/ansi.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: ansi.h,v 1.23 2011/07/17 20:54:43 joerg Exp $	*/
+/*	$NetBSD: ansi.h,v 1.24 2015/06/17 14:32:31 martin Exp $	*/
 
 /*-
  * Copyright (c) 1990, 1993
@@ -46,7 +46,7 @@
  *	#undef	_BSD_SIZE_T_
  *	#endif
  */
-#define	_BSD_CLOCK_T_		unsigned long	/* clock() */
+#define	_BSD_CLOCK_T_		unsigned int	/* clock() */
 #define	_BSD_PTRDIFF_T_		int		/* ptr1 - ptr2 */
 #define	_BSD_SIZE_T_		unsigned int	/* sizeof() */
 #define	_BSD_SSIZE_T_		int		/* byte count or error */
